+++ History.md
@@ -1,5 +1,6 @@
-## [2.11.0](https://github.com/cucumber/gherkin/compare/v2.10.0...v2.11.0)
+## [Git master](https://github.com/cucumber/gherkin/compare/v2.10.0...master)
 
+* [Java] Better exception on bad filter mix. ([#179](https://github.com/cucumber/gherkin/issues/179) Aslak Hellesøy)
 * [Java, Ruby] Formatters don't deal well with things not associated with a step ([#172](https://github.com/cucumber/gherkin/issues/172) David Kowis, Aslak Hellesøy)
 * [Java] Make model classes implement `java.io.Serializable` ([#180](https://github.com/cucumber/gherkin/issues/180) Aslak Hellesøy)